+++ src/ui/concerts/ConcertSearch.h
@@ -9,6 +9,12 @@ namespace Ui {
 class ConcertSearch;
 }
 
+namespace mediaelch {
+namespace scraper {
+class ConcertScraper;
+}
+} // namespace mediaelch
+
 class ConcertSearch : public QDialog
 {
     Q_OBJECT
@@ -21,8 +27,8 @@ public slots:
     int execWithSearch(QString searchString);
 
 public:
-    int scraperNo();
-    TmdbId scraperId();
+    mediaelch::scraper::ConcertScraper* scraper();
+    QString concertIdentifier();
     QSet<ConcertScraperInfo> infosToLoad();
 
 private: