+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 ### Added
 
  - Concert: Add IMDb/TMDb ID fields
+ - Concert search: Add language dropdown to search dialog
  - Added `*.webm` to default file filters for video files (#1122)
  - Add TVmaze as a TV scraper (#834)
    TVmaze works great for actor images and season posters.  Due to API